fix: 성향별 value int -> double

+++ projectgooroom/src/main/java/GooRoom/projectgooroom/member/domain/CleanupType.java
@@ -5,13 +5,13 @@ public enum CleanupType {
     PER_2WEEK(1),
     PER_MONTH(2);
 
-    private final int value;
+    private final double value;
 
     CleanupType(int value) {
         this.value = value;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
